Modified accuracy callouts, threat report and atis
Modified accuracy callouts so that they are more in line with F2T2EA contracts in MCM. Ranges based on TLE categories.
ATIS ends with "Lowdown [letter]" instead of "End information" which is a little to similar to a weather atis at an aerodrome :-)
Omitted "new threat active" from threat call to make it more concise. Swapped out "aircraft" for "players".
Added versioning to make it clear that this is modified.